requirements.txt
asgiref==3.6.0
black==23.3.0
boto3==1.26.105
botocore==1.29.105
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
django-jazzmin==2.6.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-stripe-subscription==1.1.13
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonfield==3.1.0
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.6
pydantic==1.10.7
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stripe==5.4.0
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
